git-svn-id: https://svn.opac.ch/svn/cr@1251 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Common.Widgets/TextFieldSlider.cs
@@ -83,7 +83,7 @@ namespace Epsitec.Common.Widgets
 			
 			if (this.Text != "")
 			{
-				this.slider.Value = (decimal) this.Value;
+				this.slider.Value = this.Value;
 			}
 		}
